+++ lib/appium_lib_core/android/device.rb
@@ -303,26 +303,30 @@ module Appium
               def start_activity(opts)
                 raise 'opts must be a hash' unless opts.is_a? Hash
 
+                option = {}
+
                 app_package = opts[:app_package]
                 raise 'app_package is required' unless app_package
 
                 app_activity = opts[:app_activity]
                 raise 'app_activity is required' unless app_activity
 
-                app_wait_package  = opts.fetch(:app_wait_package, '')
-                app_wait_activity = opts.fetch(:app_wait_activity, '')
+                option[:appPackage] = app_package
+                option[:appActivity] = app_activity
+
+                app_wait_package  = opts.fetch(:app_wait_package, nil)
+                app_wait_activity = opts.fetch(:app_wait_activity, nil)
+                option[:appWaitPackage] = app_wait_package if app_wait_package
+                option[:appWaitActivity] = app_wait_activity if app_wait_activity
 
                 unknown_opts = opts.keys - %i(app_package app_activity app_wait_package app_wait_activity)
                 raise "Unknown options #{unknown_opts}" unless unknown_opts.empty?
 
-                execute :start_activity, {}, appPackage: app_package,
-                                             appActivity: app_activity,
-                                             appWaitPackage: app_wait_package,
-                                             appWaitActivity: app_wait_activity
+                execute :start_activity, {}, option
               end
             end
 
-            # Android, Override
+            # Android, Override included method in bridge
             ::Appium::Core::Device.add_endpoint_method(:hide_keyboard) do
               def hide_keyboard(close_key = nil, strategy = nil)
                 option = {}
@@ -331,6 +335,13 @@ module Appium
                 option[:strategy] = strategy if strategy
 
                 execute :hide_keyboard, {}, option
+              end
+            end
+
+            # Android, Override included method in bridge
+            ::Appium::Core::Device.add_endpoint_method(:background_app) do
+              def background_app(duration = 0)
+                execute :background_app, {}, seconds: duration
               end
             end