+++ test/test_helper.rb
@@ -59,6 +59,10 @@ class AppiumLibCoreTest
       new.android(activity_name)
     end
 
+    def self.android_direct
+      new.android_direct
+    end
+
     def self.android_web
       new.android_web
     end
@@ -71,7 +75,7 @@ class AppiumLibCoreTest
       {
         caps: { # :desiredCapabilities is also available
           platformName: :ios,
-          automationName: 'XCUITest',
+          automationName: ENV['AUTOMATION_NAME_IOS'] || 'XCUITest',
           app: 'test/functional/app/UICatalog.app.zip',
           platformVersion: '11.4',
           deviceName: device_name,
@@ -100,7 +104,7 @@ class AppiumLibCoreTest
       {
         desired_capabilities: { # :caps is also available
           platformName: :android,
-          automationName: ENV['AUTOMATION_NAME'] || 'uiautomator2',
+          automationName: ENV['AUTOMATION_NAME_DROID'] || 'uiautomator2',
           app: 'test/functional/app/api.apk.zip',
           udid: get_udid_name,
           deviceName: 'Android Emulator',
@@ -132,12 +136,25 @@ class AppiumLibCoreTest
       }
     end
 
+    def android_direct
+      {
+        desired_capabilities: android[:desired_capabilities],
+        appium_lib: {
+          export_session: true,
+          wait: 30,
+          wait_timeout: 20,
+          wait_interval: 1,
+          direct_access: true
+        }
+      }
+    end
+
     def android_web
       {
         caps: {
           browserName: :chrome,
           platformName: :android,
-          automationName: ENV['AUTOMATION_NAME'] || 'uiautomator2',
+          automationName: ENV['AUTOMATION_NAME_DROID'] || 'uiautomator2',
           chromeOptions: { androidPackage: 'com.android.chrome', args: ['--disable-popup-blocking'] },
           # refer: https://github.com/appium/appium/blob/master/docs/en/writing-running-appium/web/chromedriver.md
           # An emulator 8.1 has Chrome/61.0.3163.98
@@ -211,7 +228,7 @@ class AppiumLibCoreTest
             warnings: {},
             desired: {
               platformName: 'Android',
-              automationName: 'uiautomator2',
+              automationName: ENV['AUTOMATION_NAME_DROID'] || 'uiautomator2',
               platformVersion: '7.1.1',
               deviceName: 'Android Emulator',
               app: '/test/apps/ApiDemos-debug.apk',
@@ -220,7 +237,7 @@ class AppiumLibCoreTest
               resetKeyboard: true
             },
             platformName: 'Android',
-            automationName: 'uiautomator2',
+            automationName: ENV['AUTOMATION_NAME_DROID'] || 'uiautomator2',
             platformVersion: '7.1.1',
             deviceName: 'emulator-5554',
             app: '/test/apps/ApiDemos-debug.apk',
@@ -259,7 +276,7 @@ class AppiumLibCoreTest
           sessionId: '1234567890',
           capabilities: {
             platformName: :android,
-            automationName: 'uiautomator2',
+            automationName: ENV['AUTOMATION_NAME_DROID'] || 'uiautomator2',
             app: 'test/functional/app/api.apk.zip',
             platformVersion: '7.1.1',
             deviceName: 'Android Emulator',
@@ -283,6 +300,44 @@ class AppiumLibCoreTest
 
       assert_requested(:post, 'http://127.0.0.1:4723/wd/hub/session', times: 1)
       assert_requested(:post, "#{SESSION}/timeouts", body: { implicit: 0 }.to_json, times: 1)
+      driver
+    end
+
+    def android_mock_create_session_w3c_direct(core)
+      response = {
+        value: {
+          sessionId: '1234567890',
+          capabilities: {
+            platformName: :android,
+            automationName: ENV['AUTOMATION_NAME_DROID'] || 'uiautomator2',
+            app: 'test/functional/app/api.apk.zip',
+            platformVersion: '7.1.1',
+            deviceName: 'Android Emulator',
+            appPackage: 'io.appium.android.apis',
+            appActivity: 'io.appium.android.apis.ApiDemos',
+            someCapability: 'some_capability',
+            unicodeKeyboard: true,
+            resetKeyboard: true,
+            directConnectProtocol: 'http',
+            directConnectHost: 'localhost',
+            directConnectPort: '8888',
+            directConnectPath: '/wd/hub'
+          }
+        }
+      }.to_json
+
+      stub_request(:post, 'http://127.0.0.1:4723/wd/hub/session')
+        .to_return(headers: HEADER, status: 200, body: response)
+
+      stub_request(:post, 'http://localhost:8888/wd/hub/session/1234567890/timeouts')
+        .with(body: { implicit: 30_000 }.to_json)
+        .to_return(headers: HEADER, status: 200, body: { value: nil }.to_json)
+
+      driver = core.start_driver
+
+      assert_requested(:post, 'http://127.0.0.1:4723/wd/hub/session', times: 1)
+      assert_requested(:post, 'http://localhost:8888/wd/hub/session/1234567890/timeouts',
+                       body: { implicit: 30_000 }.to_json, times: 1)
       driver
     end