+++ test/unit/common/websocket_test.rb
@@ -5,7 +5,7 @@ class AppiumLibCoreTest
   class WebSocketTest < Minitest::Test
     def test_connect_websocket
       ws = ::Appium::Core::WebSocket.new(url: 'ws://localhost:9292')
-      assert_equal nil, ws.client
+      assert_nil ws.client
     end
   end
 end