+++ test/unit/ios/device/mjsonwp/commands_test.rb
@@ -35,6 +35,8 @@ class AppiumLibCoreTest
           # Screen recording
 
           def test_start_recording_screen
+            skip 'Only XCUITest supports' unless @core.automation_name == :xcuitest
+
             stub_request(:post, "#{SESSION}/appium/start_recording_screen")
               .with(body: { options: { videoType: 'mjpeg', timeLimit: '180', videoQuality: 'medium' } }.to_json)
               .to_return(headers: HEADER, status: 200, body: { value: ['a'] }.to_json)
@@ -45,6 +47,8 @@ class AppiumLibCoreTest
           end
 
           def test_start_recording_screen_custom
+            skip 'Only XCUITest supports' unless @core.automation_name == :xcuitest
+
             stub_request(:post, "#{SESSION}/appium/start_recording_screen")
               .with(body: { options: {
                 videoType: 'libx264', timeLimit: '60', videoQuality: 'medium', videoScale: '320:240'
@@ -57,6 +61,8 @@ class AppiumLibCoreTest
           end
 
           def test_start_recording_screen_custom_force
+            skip 'Only XCUITest supports' unless @core.automation_name == :xcuitest
+
             stub_request(:post, "#{SESSION}/appium/start_recording_screen")
               .with(body:
                 { options: { forceRestart: true, videoType: 'libx264', timeLimit: '60', videoQuality: 'medium' } }.to_json)
@@ -68,6 +74,8 @@ class AppiumLibCoreTest
           end
 
           def test_stop_recording_screen_default
+            skip 'Only XCUITest supports' unless @core.automation_name == :xcuitest
+
             stub_request(:post, "#{SESSION}/appium/stop_recording_screen")
               .with(body: {}.to_json)
               .to_return(headers: HEADER, status: 200, body: { value: ['a'] }.to_json)
@@ -89,6 +97,8 @@ class AppiumLibCoreTest
           end
 
           def test_get_battery_info
+            skip 'Only XCUITest supports' unless @core.automation_name == :xcuitest
+
             stub_request(:post, "#{SESSION}/execute")
               .with(body: { script: 'mobile: batteryInfo', args: [{}] }.to_json)
               .to_return(headers: HEADER, status: 200, body: { value: { state: 3, level: 1.0 } }.to_json)
@@ -98,6 +108,32 @@ class AppiumLibCoreTest
             assert_requested(:post, "#{SESSION}/execute", times: 1)
             assert_equal :full, info[:state]
             assert_equal 1.0, info[:level]
+          end
+
+          def test_method_missing
+            stub_request(:get, "#{SESSION}/element/id/attribute/name")
+              .to_return(headers: HEADER, status: 200, body: { value: '' }.to_json)
+
+            e = ::Selenium::WebDriver::Element.new(@driver.send(:bridge), 'id')
+            e.name
+
+            assert_requested(:get, "#{SESSION}/element/id/attribute/name", times: 1)
+          end
+
+          def test_background_app
+            if @core.automation_name == :xcuitest
+              stub_request(:post, "#{SESSION}/appium/app/background")
+                .with(body: { seconds: { timeout: 0 } }.to_json)
+                .to_return(headers: HEADER, status: 200, body: { value: '' }.to_json)
+            else
+              stub_request(:post, "#{SESSION}/appium/app/background")
+                .with(body: { seconds: 0 }.to_json)
+                .to_return(headers: HEADER, status: 200, body: { value: '' }.to_json)
+            end
+
+            @driver.background_app 0
+
+            assert_requested(:post, "#{SESSION}/appium/app/background", times: 1)
           end
         end # class CommandsTest
       end # module MJSONWP